requirements.txt
feedparser
moviepy
Pillow
openai
# gTTS
# langchain
# langchain-openai
# langchain-community
dotenv
ffmpeg-python
flask